% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignTF_SNP.R
\name{alignTF_SNP}
\alias{alignTF_SNP}
\title{Align TF around a SNP of interest}
\usage{
alignTF_SNP(list, PFMatrixList)
}
\arguments{
\item{list}{Data.frame outputed from getCandidateTFs().}

\item{PFMatrixList}{Output of getMotifMatrixJASPAR() using matrix="PFM", with all the TF you want to query.}
}
\description{
By using the PWM and the genomic locations of the motif found, it adds blanks to the PWM to make them be centered in the SNP of interest.
}
