% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMotifsRegions.R
\name{annotateMotifsRegions}
\alias{annotateMotifsRegions}
\title{Annotate motifs in regions of interest}
\usage{
annotateMotifsRegions(bed, genome = "hg19", motif_file,
  out_name = "motifs_genome.bed", path_homer = "default")
}
\arguments{
\item{bed}{Bed file containing the regions you want to analyze.}

\item{genome}{Name of the reference genome to use. "hg19" is set as default.}

\item{motif_file}{Character with the path and name of the motif_file generated by 
catSignMotifsHOMER().}

\item{out_name}{Name of the output file.}

\item{path_homer}{Path for homer location in your computer. It automatically looks in the aliases.}
}
\value{
It will return two files: output_file.bed, which wil contain the coordinates for the motifs analyzed and
output_file.tsv, which will contain the annotation information and the data of which motif was found in which region.
}
\description{
Looks at instances of motifs of interest along the entire genome.
}
\examples{
\dontrun{
annotateMotifsRegions(bed="regions.bed", motif_file="allMotifs.motif", out_name="myOutput")
}
}
