% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARGE_package.R
\name{backward_sel_WIC}
\alias{backward_sel_WIC}
\title{backward_sel_WIC}
\usage{
backward_sel_WIC(Y, N, n, B_new, id, family = "gaussian",
  corstr = "independence", nb = F, is.gee = F, ...)
}
\arguments{
\item{Y}{: the response variable.}

\item{N}{: the number of clusters.}

\item{n}{: the cluster size.}

\item{B_new}{: the model matrix.}

\item{id}{: the ID for each individual in the cluster.}

\item{family}{: the specified "exponential" family for the GLM. The default is \code{family="gaussian"}.}

\item{corstr}{:the specified "working correlation" structure. The default is \code{corstr="independence"}.}

\item{nb}{: a logical argument, is the model a negative binomial model? The default is \code{FALSE}.}

\item{is.gee}{: is this a GEE model? The default is \code{FALSE}.}

\item{...}{: further arguments passed to or from other methods.}
}
\value{
\code{backward_sel_WIC} returns the Wald statistic from the fitted model (the penalty is apllied later on).
}
\description{
Backward selection funciton for MARGE - uses the Wald information criterion (WIC).
}
\references{
Stoklosa, J. Gibb, H. Warton, D.I. Fast forward selection for Generalized Estimating Equations With a Large Number of Predictor Variables. \emph{Biometrics}, 70, 110-120.

Stoklosa, J. and Warton, D.I. (2017). A generalized estimating equation approach to multivariate adaptive regression splines. \emph{Journal of Computational and Graphical Statistics}, in review.
}
\seealso{
\code{\link{backward_sel}}
}
\author{
Jakub Stoklosa and David I. Warton
}
