% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catSignMotifsHOMER.R
\name{catSignMotifsHOMER}
\alias{catSignMotifsHOMER}
\title{Join significant de novo motifs found by HOMER}
\usage{
catSignMotifsHOMER(path_output, num_sign)
}
\arguments{
\item{path_output}{Path of the output folder that was generated containing HOMER reports.}

\item{num_sign}{Number of significant motifs that were found.}
}
\value{
A file "path_output.motif" with all the motifs found significant in the regions.
}
\description{
Creates a unique .motif file containing all the significant motifs that were found
by HOMER de novo motif analysis.
}
\examples{
\dontrun{
catSignMotifsHOMER("results_homer")
}
}
