% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_homer.R
\name{check_homer}
\alias{check_homer}
\title{Check for working HOMER installation}
\usage{
check_homer()
}
\value{
logical; \code{TRUE} for success, \code{FALSE} otherwise
}
\description{
Used to check that HOMER is installed and working on package attach,
and reads/sets the \code{options('homer_path')} if user has not.
}
