% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSignHOMER.R
\name{countSignHOMER}
\alias{countSignHOMER}
\title{Count significant results from de Novo motif analysis}
\usage{
countSignHOMER(results_homer)
}
\arguments{
\item{results_homer}{Either the path of the output folder that was generated containing HOMER reports or
the data.frame obtained using XML::readHTMLTable.}
}
\value{
The number of significant motifs (int).
}
\description{
Reads the html output from homer and counts how many motifs were classified as significant.
}
\examples{
\dontrun{
num <- countSignHOMER("results_homer")
}
}
