% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deNovoToDataFrame.R
\name{deNovoToDataFrame}
\alias{deNovoToDataFrame}
\title{Creates data.frame with de novo HOMER results}
\usage{
deNovoToDataFrame(results_homer, include_all = TRUE)
}
\arguments{
\item{results_homer}{Character string containing the path of the output folder that was generated 
with HOMER.}

\item{include_all}{TRUE (default) of FALSE for including also those motifs that have no ensembl_gene_id.}
}
\value{
Data.frame containing the results and gene annotation for the de novo motifs.
}
\description{
Generates a data.frame containing the significant motifs found by HOMER.
}
\examples{
\dontrun{
df.tf <- deNovoToDataFrame("results_homer", include_all=FALSE)
}
}
