% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictHomerMotifs.R
\docType{data}
\name{dictHomerMotifs}
\alias{dictHomerMotifs}
\title{Dictionary of Homer Motifs}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Motif.Name}{Name of the motif, as outputed by HOMER.}
  \item{Motif.Symbol}{Section of the Motif.Name including the symbol.}
  \item{external_gene_name}{Standarized gene symbol.}
  \item{ensembl_gene_id}{ENSEMBL gene id.}
}}
\source{
\url{HOMER}
}
\usage{
dictHomerMotifs
}
\description{
A dataset containing the names of the motifs used by HOMER and their 
corresponding symbols and ensembl_gene_id. Useful when trying to 
intersect HOMER motif data with transcription data.#'
}
\keyword{datasets}
