% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{.parse_homer_subfields}
\alias{.parse_homer_subfields}
\title{Parse HOMER subfields}
\usage{
.parse_homer_subfields(motif_tbl)
}
\arguments{
\item{motif_tbl}{a motif_tbl constructed while running \code{read_*_results}
and consequently \code{read_motif}, or a custom tibble that
has the above field separators as described with columns
\code{motif_name} and \code{experiment}}
}
\value{
Adds the above additional columns to an existing motif
  tibble returned by \code{read_motif} or \code{read_denovo_results}
}
\description{
An internal function that parses the \code{motif_name} and
\code{experiment} subfields present in HOMER-formatted motifs.
}
\details{
HOMER-formatted motifs contain three key fields:

\itemize{
  \item \code{motif_name} human-readable name
  \item \code{experiment} how the motif was identified
  \item \code{database} what database the motif is from
}

Furthermore, each motif may contain additional information enclosed
in parentheses. This includes:

\itemize{
  \item \code{motif_family} included in the \code{motif_name} field,
    describes the transcription factor family.
  \item \code{accession} included in the \code{experiment} field,
    provides the specific ID or publication of the raw/processed data.
}
}
\seealso{
\code{\link{read_motif}}
}
