% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawAlignedMotifSNP.R
\name{drawAlignedMotifSNP}
\alias{drawAlignedMotifSNP}
\title{Draw list of motif logos in stack}
\usage{
drawAlignedMotifSNP(mot.al.pos, length, posSNP)
}
\arguments{
\item{length}{Total length of the motifs that are being plotted.}

\item{posSNP}{Integer with the position from the start of the SNP of interest.}

\item{list}{List of pfm objects. Also output from alignTF_SNP().}
}
\description{
Using a list of pfm objects, it draws a list of motifs and highlights the position where the SNP can be found.
}
