% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_homer.R
\name{get_homer_bin}
\alias{get_homer_bin}
\title{Get the base installation directory of HOMER}
\usage{
get_homer_bin()
}
\value{
character vector with base directory of HOMER installation
}
\description{
Grabs the base directory of HOMER installation from \code{options('homer_path')};
the correct base path should be set with package attachment, or alert user
to check/set the path manually. This function is kept to keep code from breaking.
}
