% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCandidateTFs.R
\name{getCandidateTFs}
\alias{getCandidateTFs}
\title{Obtain candidate TF that might be binding a specified region}
\usage{
getCandidateTFs(dat, PFMatrixList, type = "region", seqname,
  minscore = "80\%", strand = "*", start = NA, chr = NA)
}
\arguments{
\item{dat}{Object of type GRanges or data.frame for genomic region (it will only use the first row of these objects), or character or DNAstring for type sequence.}

\item{PFMatrixList}{Object of type "PFMatrixList" containing the subset of motif matrices against whic compare our sequence.}

\item{seqname}{Character string with the name of the region.}

\item{minscore}{Minimum percentage for matching motifs. Default: "80%".}

\item{strand}{String containing the strand in which to perform the analysis. "*" both strands (default), "+" positive strand and "-" negative strand.}

\item{start}{Integer containing the starting coordinates of region, if not provided using a data.frame or GRanges object.}

\item{region}{Character string with the type of dat you are using as input: "region" (default) for using genomic coordinates as input or "seq" for using a nucleotide sequence as input.}

\item{end}{Character string containing chromosome name, if not provided using a data.frame or GRanges object.}
}
\value{
Data.frame containing all the TF of interest that can be found in the region.
}
\description{
Using an input DNA sequence or genomic coordinates, it returns a list of candidate TF that match motif sequences found in your region.
}
\examples{
\\notrun{

}
}
