% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifMatrixJASPAR.R
\name{getMotifMatrixJASPAR}
\alias{getMotifMatrixJASPAR}
\title{Get JASPAR2016 Motif Matrix}
\usage{
getMotifMatrixJASPAR(species = 9606, matrixtype = "PWM")
}
\arguments{
\item{species}{Integer containing the taxa id for the sepecie you want to query. Default 9606 (Homo sapiens).}

\item{matrixtype}{Character containing the type of motif matrix you want to retrieve from JASPAR database. One of "PWM" (default), "ICM", "PFM".}
}
\value{
Object of type "PFMatrixList" that can be used as input for getCandidateTFs.
}
\description{
Obtains a PWMatrixList object containing all the PWM that match the query generated to JASPAR2016 database.
}
\examples{
\\notrun{
getMotifMatrixJASPAR(species=9606, matrixtype="PWM")
}
}
