% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNameFromAlias.R
\name{getNameFromAlias}
\alias{getNameFromAlias}
\title{Get gene symbol from alias.}
\usage{
getNameFromAlias(alias)
}
\arguments{
\item{alias}{Character vector containing the aliases for which you want to obtain the gene symbols.}
}
\value{
Data.frame with the gene symbol and the alias.
}
\description{
Obtains a the gene symbol that corresponds to a gene alias.
}
\examples{
\dontrun{
geneNames <- getNameFromAlias(alias)
}
}
