% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARGE_package.R
\name{getNumberPart}
\alias{getNumberPart}
\title{getNumberPart}
\usage{
getNumberPart(x)
}
\arguments{
\item{x}{: a numerical value.}
}
\description{
A function to pull out negative and decimals from strings.
}
