% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignNamesHOMER.R
\name{getSignNamesHOMER}
\alias{getSignNamesHOMER}
\title{Obtain names from HOMER TF motif names.}
\usage{
getSignNamesHOMER(results_homer, num_sign)
}
\arguments{
\item{results_homer}{Either the path of the output folder that was generated containing HOMER reports or
the data.frame obtained using XML::readHTMLTable.}

\item{num_sign}{Number of significant motifs that were found.}
}
\value{
A vector with the names of the TF.
}
\description{
Obtains human readable names (SYMBOLS) from HOMER de novo motif name matching.
}
\examples{
\dontrun{
getSignNamesHOMER("results_homer", 2)
}
}
