% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_homer.R
\name{list_homer_packages}
\alias{list_homer_packages}
\title{List installed/available HOMER packages}
\usage{
list_homer_packages(installed_only = TRUE)
}
\arguments{
\item{installed_only}{only show installed packages (default: TRUE)}
}
\value{
\code{tibble} with status of installed (+) or not (-),
  the package, version, and description.
}
\description{
Runs the configureHomer.pl script to get list of installed and
available packages, parsing the output into a tidy tibble.
}
