% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/marge.R
\name{marge.og}
\alias{marge.og}
\title{original Marge for R}
\usage{
marge.og(...)
}
\description{
Deprecated. A function for people who type too fast for their own good and perform a lot
of merges in R.
}

