% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.marge.R
\name{print.marge}
\alias{print.marge}
\title{Print for objects of class 'marge'}
\usage{
\method{print}{marge}(x, pen = c("2", "logN"), ...)
}
\arguments{
\item{x}{a marge model object}

\item{pen}{a set penalty used for the GCV (which model information to return). The default is "2".}

\item{...}{}
}
\value{
printed argument
}
\description{
Print for objects of class 'marge'
}
\examples{
#' data(leptrine)
dat1 <- leptrine[[1]]
mod <- marge(Y ~ RAIN_DRY_QTR + FC, data = dat1, family = "binomial", N = nrow(dat1), n = 1)
mod
}
