% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_denovo_results}
\alias{read_denovo_results}
\title{Read Denovo Enriched Motifs HOMER Output}
\usage{
read_denovo_results(path, homer_dir = TRUE)
}
\arguments{
\item{path}{path to the HOMER directory where all outputs are
stored}

\item{homer_dir}{does the path point to a HOMER directory;
if \code{FALSE}, path must point to the file directly
[default: TRUE]}
}
\value{
a tibble with the following columns:
\itemize{
  \item \code{consensus} the consensus sequence of the denovo motif
  \item \code{motif_name} name of the motif
  \item \code{log_odds_detection} threshold used to determine bound vs. unbound sites
  \item \code{motif_pwm} a list column with PWMs for each motif
  \item \code{log_p_value_detection} from the original experiment used to ID motif
  \item \code{tgt_num} number of times motif appears in target sequences
  \item \code{tgt_pct} percent of times motif appears in target sequences
  \item \code{bgd_num} number of times motif appears in background sequences
  \item \code{bgd_pct} percent of times motif appears in background sequences
  \item \code{log_p_value} final enrichment from experiment -log10(p-value)
  \item \code{tgt_pos} average position of motif in target sequences, where
    0 = start of sequences
  \item \code{tgt_std} standard deviation of position in target sequences
  \item \code{bgd_pos} average position of motif in background sequences,
    where 0 = start of sequences
  \item \code{bgd_std} standard deviation of position in background sequences
  \item \code{strand_bias} log ratio of + strand occurrences to - strand occurrences
  \item \code{multiplicity} average number of occurrences per sequence in
    sequences with 1 or more binding sites
}
}
\description{
Reads in results from a denovo motif enrichment analysis created by
HOMER.
}
\details{
Following an analysis using \code{\link{find_motifs_genome}}, a HOMER directory
is created which analyses the enrichment of denovo motifs. This
function reads and parses in that file into a tidy format.
}
\seealso{
\code{\link{read_known_results}}, \code{\link{find_motifs_genome}}
}
