% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_motif}
\alias{read_motif}
\title{Read HOMER Motif Files}
\usage{
read_motif(path)
}
\arguments{
\item{path}{location of motif file}
}
\value{
at minimum, a tibble with the following columns:
\itemize{
  \item \code{consensus} the consensus sequence of the denovo motif
  \item \code{motif_name} name of the motif
  \item \code{log_odds_detection} threshold used to determine bound vs. unbound sites
  \item \code{motif_pwm} a list column with PWMs for each motif
}

The following columns are presented when available from complete \code{*.motif*} files
or from HOMER results directories:

\itemize{
  \item \code{log_p_value_detection} from the original experiment used to ID motif
  \item \code{tgt_num} number of times motif appears in target sequences
  \item \code{tgt_pct} percent of times motif appears in target sequences
  \item \code{bgd_num} number of times motif appears in background sequences
  \item \code{bgd_pct} percent of times motif appears in background sequences
  \item \code{log_p_value} final enrichment from experiment -log10(p-value)
  \item \code{tgt_pos} average position of motif in target sequences, where
    0 = start of sequences
  \item \code{tgt_std} standard deviation of position in target sequences
  \item \code{bgd_pos} average position of motif in background sequences,
    where 0 = start of sequences
  \item \code{bgd_std} standard deviation of position in background sequences
  \item \code{strand_bias} log ratio of + strand occurrences to - strand occurrences
  \item \code{multiplicity} average number of occurrences per sequence in
    sequences with 1 or more binding sites
}
}
\description{
The core function to read in motif files, whether from the HOMER database,
from HOMER denovo motif enrichment results, or even custom motifs. In all
cases, these files must be in the HOMER-format. See below for more details.
}
\details{
To read-in a HOMER-formatted motif, at a minimum, the first three
fields are required to properly ID the motif:

\itemize{
  \item \code{">" + Consensus sequence} The dominant or likeliest sequence
  \item \code{Motif name} Should be unique
  \item \code{Log odds detection threshold} determines bound vs. unbound sites
}

The remaining extra fields of HOMER-formatted motifs are described at the
URL below, and primarily meant for interpreting motifs from HOMER's
own database. To read more about the HOMER format, see:
\url{http://homer.ucsd.edu/homer/motif/creatingCustomMotifs.html}

Note that HOMER also has additional information in the motif name
regarding its origin and identity. See the internal function \code{.parse_homer_subfields}
for more info and to break this field up.

Subsequent lines (after the ">") describe the position weight matrix (PWM),
with columns in order of A, C, G, T describing the probabilities of
per position of each nucleotide.

Note that it is possible to combine complete information (HOMER-formatted) motifs
with minimal motifs. Simply use \code{dplyr::bind_rows} for easy concatenation
despite column spec differences.
}
