% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARGE_package.R
\name{sand_fun}
\alias{sand_fun}
\title{sand_fun}
\usage{
sand_fun(Y, X, N, n_vec, mu.est, V.est, nb = T, omega = 1, ...)
}
\arguments{
\item{Y}{: the response variable.}

\item{X}{: the model matrix.}

\item{N}{: the number of clusters.}

\item{n_vec}{: a vector consisting of the number of cluster sizes.}

\item{mu.est}{: esitmates of the fitted mean function under the null model.}

\item{V.est}{: esitmates of the fitted variance function under the null model.}

\item{nb}{: a logical argument, is the model a negative binomial model? The default is \code{FALSE}.}

\item{omega}{: a regulariztion constant that is applied if your data is of high dimension (n>N). The default is \code{omega=1}.}

\item{...}{: further arguments passed to or from other methods.}
}
\value{
\code{sand_fun} returns the sandwich (or Pan's) standard error using estimates from the fitted GEE under an independent working correlation.
}
\description{
A function that calculates the sandwich (or Pan's) standard error using estimates from the fitted GEE under an independent working correlation.
}
\details{
This function was used for the Arthropod example.
}
\references{
Liang, K. Y. and Zeger, S. L. (1986). Longitudinal aata analysis using generalized linear models. \emph{Biometrika}, 73, 13-22.

Stoklosa, J., Gibb, H. and Warton, D.I. (2014). Fast forward selection for generalized estimating equations with a large number of predictor variables. \emph{Biometrics}, 70, 110-120.

Stoklosa, J. and Warton, D.I. (2018). A generalized estimating equation approach to multivariate adaptive regression splines. \emph{Journal of Computational and Graphical Statistics}, \strong{27}, pp. 245--253.
}
\author{
Jakub Stoklosa and David I. Warton.
}
