% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARGE_package.R
\name{stat_out_score_glm_null}
\alias{stat_out_score_glm_null}
\title{stat_out_score_glm_null}
\usage{
stat_out_score_glm_null(Y, family = "gaussian", B_null, nb = F, ...)
}
\arguments{
\item{Y}{: the response variable.}

\item{family}{: the specified "exponential" family for the GLM. The default is \code{family = "gaussian"}.}

\item{B_null}{: model matrix under the null model.}

\item{nb}{: a logical argument, is the model a negative binomial model? The default is \code{FALSE}.}

\item{...}{: further arguments passed to or from other methods.}
}
\value{
\code{stat_out_score_glm_null} returns a list of values (mainly products of matrices) that make up the final score statistic calculation (required for another function).
}
\description{
A function that calculates parts of the score statistic for GLMs only (it is used for the full path for forward selection).
}
\references{
Stoklosa, J., Gibb, H. and Warton, D.I. (2014). Fast forward selection for generalized estimating equations with a large number of predictor variables. \emph{Biometrics}, \strong{70}, 110--120.

Stoklosa, J. and Warton, D.I. (2018). A generalized estimating equation approach to multivariate adaptive regression splines. \emph{Journal of Computational and Graphical Statistics}, \strong{27}, 245--253.
}
\seealso{
\code{\link{stat_out}} and \code{\link{stat_out_score_glm_null}}
}
\author{
Jakub Stoklosa and David I. Warton.
}
