% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARGE_package.R
\name{stat_out_score_null}
\alias{stat_out_score_null}
\title{stat_out_score_null}
\usage{
stat_out_score_null(Y, N, n, id, family = "gaussian",
  corstr = "independence", B_null, nb = F, is.gee = F, ...)
}
\arguments{
\item{Y}{: the response variable.}

\item{N}{: the number of clusters.}

\item{n}{: the cluster size.}

\item{id}{: the ID for each individual in the cluster.}

\item{family}{: the specified "exponential" family for GLMs. The default is \code{family="gaussian"}.}

\item{corstr}{: the specified "working correlation" structure. The default is \code{corstr="independence"}.}

\item{B_null}{: model matrix under the null model.}

\item{nb}{: a logical argument, is the model a negative binomial model? The default is \code{FALSE}.}

\item{is.gee}{: a logical argument, is this a GEE model? The default is \code{FALSE}.}

\item{...}{: further arguments passed to or from other methods.}
}
\value{
\code{stat_out_score_null} returns a list of values (mainly products of matrices) that make up the final score statistic calculation (required for another function).
}
\description{
A function that calculates parts for the score statistic for GEEs (it is used for the full path for forward selection).
}
\details{
The null model used here is by definition the current parent model. We compare the alternative model (a new basis function combined with the parent) with the null model. In the code used, only the null model is fit (the dispersion parameter is also estimated for GEE) and then the score statistic is obtained.
}
\references{
Stoklosa, J., Gibb, H. and Warton, D.I. (2014). Fast forward selection for generalized estimating equations with a large number of predictor variables. \emph{Biometrics}, 70, 110-120.

Stoklosa, J. and Warton, D.I. (2018). A generalized estimating equation approach to multivariate adaptive regression splines. \emph{Journal of Computational and Graphical Statistics}, \strong{27}, pp. 245--253.
}
\seealso{
\code{\link{stat_out}} and \code{\link{stat_out_score_glm_null}}
}
\author{
Jakub Stoklosa and David I. Warton.
}
