% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pwm.R
\name{write_homer_motif}
\alias{write_homer_motif}
\title{Write Motif PWM in HOMER format}
\usage{
write_homer_motif(
  motif_pwm,
  motif_name,
  log_odds_detection,
  consensus = NULL,
  file,
  append = TRUE
)
}
\arguments{
\item{motif_pwm}{tibble with alphabet columns}

\item{motif_name}{name of the sequence}

\item{log_odds_detection}{the threshold for detecting given motif}

\item{consensus}{the accepted consensus motif sequence
[optional: calculated from pwm if not provided]}

\item{file}{path to new motif file; suffix should be '.motif'}

\item{append}{whether to overwrite existing file [default: TRUE, e.g.
add on to existing file]}
}
\value{
nothing; called for its side-effect of producing HOMER motif file
}
\description{
Create a minimal HOMER formatted motif file based on PWM and required
motif info.
}
\examples{
\dontrun{
knw_pwm <- read_known_results("inst/extdata/knownResults.txt", homer_dir = FALSE)

## Write multiple motifs
pwalk(list(motif_pwm = knw_pwm$motif_pwm,
           motif_name = knw_pwm$motif_name,
           log_odds_detection = knw_pwm$log_odds_detection),
      write_homer_motif, file = "test.motif", append = TRUE)
}
}
