% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deNovoToDataFrame.R
\name{deNovoToDataFrame}
\alias{deNovoToDataFrame}
\title{Creates data.frame with de novo HOMER results}
\usage{
deNovoToDataFrame(results_homer)
}
\arguments{
\item{results_homer}{Either the path of the output folder that was generated containing HOMER reports or
the data.frame obtained using XML::readHTMLTable.}
}
\value{
The number of significant motifs (int).
}
\description{
Generates a data.frame containing the significant motifs found by HOMER.
}
\examples{
\dontrun{
num <- countSignHOMER("results_homer")
}
}
