% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/marge.R
\name{marge}
\alias{marge}
\title{Marge for R}
\usage{
marge(speak = "Shut up, Becky!", ...)
}
\arguments{
\item{speak}{What should Marge say? Defaults to "Shut up, Becky!"}
}
\description{
A function for people who type too fast for their own good and perform a lot
of merges in R.
}

