% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_denovo_html_results.R
\name{read_denovo_html_results}
\alias{read_denovo_html_results}
\title{Read de novo enriched motifs from HOMER .info.html output files}
\usage{
read_denovo_html_results(path, homer_dir = TRUE)
}
\arguments{
\item{path}{path to the HOMER directory where all outputs are
stored}

\item{homer_dir}{does the path point to a HOMER directory;
if \code{FALSE}, path must point to the file directly
[default: TRUE]}
}
\value{
a list with the following structure:
\itemize{
  \item \code{Motif_information} a dataframe with most of the info returned by \code{\link{read_denovo_results}}
  \item \code{Matches_to_known_motifs} Dataframe with information on related known motifs matched to the de novo motif
}
The \code{Motif_information} dataframe has the following columns:
\itemize{
  \item \code{motif_name} name of the motif
  \item \code{consensus} the consensus sequence of the denovo motif
  \item \code{p_value} final enrichment from experiment (p-value)
  \item \code{log_p_value} final enrichment from experiment -log10(p-value)
  \item \code{info_content} information Content per bp
  \item \code{tgt_num} number of times motif appears in target sequences
  \item \code{tgt_pct} percent of times motif appears in target sequences
  \item \code{bgd_num} number of times motif appears in background sequences
  \item \code{bgd_pct} percent of times motif appears in background sequences
  \item \code{tgt_pos} average position of motif in target sequences, where 0 = start of sequences
  \item \code{bgd_pos} average position of motif in background sequences
  \item \code{strand_bias} strand Bias (log2 ratio + to - strand density)
  \item \code{multiplicity} multiplicity (# of sites on avg that occur together)
}
#' The \code{Matches_to_known_motifs} dataframe has the following columns:
\itemize{
  \item \code{motif_name} name of the motif
  \item \code{motif_family} family of the motif
  \item \code{ID} ID of the motif or the experiment
  \item \code{database} database of the motif
  \item \code{rank} match rank
  \item \code{score} log odds score of the motif matrix, higher scores are better matches
  \item \code{offset} from the center of the region
  \item \code{orientation} forward or reverse (on average?)
  \item \code{original_alignment} consensus sequence for comparison
  \item \code{matched_alignment} matched sequence with mismatches in "-" from the consensus. Can be offset.
}
#' @importFrom stringr str_replace str_split str_detect
}
\description{
Reads in results from all individual info.html files produced by HOMER de novo motif enrichment analysis.
}
\details{
Following an analysis using \code{\link{find_motifs_genome}}, a HOMER directory
is created which analyses the enrichment of known motifs. This
function reads and parses in the .html file into a tidy format.

It might be more useful to read in the .html files instead of the .motifs file because
the .html contain information on known similar motifs.
}
