% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.marge.R
\name{summary.marge}
\alias{summary.marge}
\title{Summary for objects of class 'marge'}
\usage{
\method{summary}{marge}(object, ...)
}
\arguments{
\item{object}{a marge model object}

\item{...}{}

\item{pen}{a set penalty used for the GCV (which model information to return). The default is "2".}
}
\value{
printed argument
}
\description{
Summary for objects of class 'marge'
}
\examples{
data(leptrine)
dat1 <- leptrine[[1]]
mod <- marge(Y ~ RAIN_DRY_QTR + FC, data = dat1, family = "binomial", N = nrow(dat1), n = 1)
summary(mod)
}
